+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2.11.12
+
+- `Fix` — Make Toolbox tooltip position font-size independent  
+
+### 2.11.11
+
+- `New` — Add ability to pass configuration for internal Tools
+
 ### 2.11.10
 
 - `Fix` - Fix editor view on mobile devices